Extract a separate class to define address submit.
